requirements.txt
einops
loguru
tokenizers
fire
pytorch-ignite
accelerate
transformers
bitsandbytes
scipy
webdataset
phonemizer
numpy
pyyaml
pandas 
setuptools
soundfile
fairseq2
sonar-space
sentence_transformers
torch>=1.8.2
torchaudio
torchinfo
